requirements.txt
arrow==1.3.0
asgiref==3.9.1
binaryornot==0.4.4
build==1.2.2.post1
CacheControl==0.14.1
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
cookiecutter==2.6.0
crashtest==0.4.1
distlib==0.3.9
dj-database-url==3.0.1
Django==5.2.4
django-browser-reload==1.19.0
django-tailwind==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dulwich==0.21.7
fastjsonschema==2.21.1
filelock==3.16.1
flake8==7.1.2
gunicorn==23.0.0
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.6
keyring==24.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
numpy==2.3.5
packaging==24.2
pexpect==4.9.0
pillow==11.3.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
psycopg==3.2.10
psycopg-binary==3.2.10
ptyprocess==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-slugify==8.0.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
types-python-dateutil==2.9.0.20250708
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.28.0
whitenoise==6.6.0
